Add module and use arrow function
